Leverage iam modules to create groups, policies, and roles
Each of these currently has their own remote state store. This may be too
fine-grained, but for now I find it makes the most sense for organization.

It is also worth noting that my intention for these resources are "global"
level. Esspecially for policies and roles, it probably makes more sense to
co-locate these resources next to dependent resource. E.G. a bastion instance
should probably define its instance profile (with role and policy) within the
same module as the EC2 instance.

However, to be honost, I still am not quite sure on successful patterns for how
to organize all of these configuration files and state.